Added a clarification
Added a clarification for when we get the following error from MS:

"2148916238: The account is a child (under 18) and cannot proceed unless the account is added to a Family by an adult. This only seems to occur when using a custom Microsoft Azure application. When using the Minecraft launchers client id, this doesn't trigger. Source: https://wiki.vg/Microsoft_Authentication_Scheme"

Indicating that it is an issue with the way your login token is configured and not the user's account.