Demo account and logging changes
Logs produced by the raw input method now properly report that they're from msmc

Updated the logging tag msmc uses to be more inline with mclc

The login script no longer fails for a demo account. If used with mclc the --demo launch parameter is added automatically if appropriate. There is a new function "isDemoUser" for any other launcher frameworks

And due to the votes we got back. The player name when the game is launched in demo mode is now "Player". Check the discord for the poll.