Friend lists!
Added friends lists to the object returned by the 'getXbox' function. Was trying to add gamepass...but ended up finding a method to pull a user's xbox friend-list.

We'll continue to look for ways to add gamepass detection, but at this moment that seems unlikely to materialise before the final release.